Various fixes and improvements

+++ classes/task/emit_task.php
@@ -25,7 +25,6 @@
 
 namespace logstore_caliper\task;
 
-use tool_log\log\manager;
 use logstore_caliper\log\store;
 
 defined('MOODLE_INTERNAL') || die();
@@ -52,6 +51,10 @@ class emit_task extends \core\task\scheduled_task {
         $store = new store($manager);
 
         $events = $DB->get_records('logstore_caliper_log');
+        if (empty($events)) {
+            return;
+        }
+
         $store->process_events($events);
 
         $DB->delete_records_list('logstore_caliper_log', 'id', array_keys($events));
